Correct lifetime for callbacks
Stores the callbacks (for notifications and input both) in boxes, so
both the function and its destructor can be type-erased. Also changes
the basic callback type to FnMut(&Data) + Send + 'static, to indicate
that the callback can be called from a different thread and after the
function registering the callback returns, and that the callback
doesn't get to hold any references to the data.

Also changes the signature of Packet to just yield a data slice; the
old approach of returning an iterator was unsafe because the iterator
could outlive the data pointer. Some packet types get lifetimes and
PhantomData enforcing those lifetimes as well.

Increments the version to 0.2.0, is the change to the callback and
packet types is clearly breaking from a semver perspective.

Fixes #2