Gave test data the proper retain count
In the test data for notification_from_property_changed, the CFString that was wrapping the CFStringRef for the .propertyName was releasing the CFStringRef at the end of notification_raw's instantiation (to a retain count of 0), so the CFStringRef that was returned was invalid. Instead, the CFString needs to live until after the call Notification::from has returned so its CFStringRef can be valid for the whole call.