Fixing SIGILL in tests caused by over-release of CFStringRef
Because the MIDINotification struct is not owned by the callback function (it's just a pointer to the data), the CFStringRef that it contains comes into the callback with a reference count of 1. Therefore we should follow CoreFoundation's "The Get Rule" and increase the reference count before it is decreased again when the CFString is dropped after being converted to a Rust String. Otherwise the retain count goes below 0 as the test function ends, causing a panic.